backend/requirements.txt
flask
supabase
python-dotenv
Flask-Cors 